refactor: consolidate stock logic into server index
